commit for merge

+++ graphql/resolvers/FlightInstance.js
@@ -54,13 +54,13 @@ const FlightInstance = {
       .then(constructed => {
         return db.FlightInstance.findById(FlightInstanceId)
         .then(foundInstance => {
-          return foundInstance.update(instanceUpdates)
+          return foundInstance.update(constructed)
         })
       })
       .then(updatedInstance => {
         // add or remove attachments
         var attachmentsPromiseArr = []
-        if (data.addAttachments.length) {
+        if (data.addAttachments && data.addAttachments.length) {
           data.addAttachments.forEach(attachment => {
             var attachmentPromise = db.Attachment.create({
               FlightInstanceId: FlightInstanceId,
@@ -73,7 +73,7 @@ const FlightInstance = {
             attachmentsPromiseArr.push(attachmentPromise)
           })
         }
-        if (data.removeAttachments.length) {
+        if (data.removeAttachments && data.removeAttachments.length) {
           data.removeAttachments.forEach(id => {
             var attachmentPromise = db.Attachment.destroy({where: {id: id}})
             attachmentsPromiseArr.push(attachmentPromise)
